+++ lib/consistent_random.rb
@@ -1,24 +1,55 @@
 # frozen_string_literal: true
 
-require_relative "consistent_random/context"
+require "digest/sha1"
+require "securerandom"
+
 require_relative "consistent_random/rack_middleware"
 require_relative "consistent_random/sidekiq_middleware"
+require_relative "consistent_random/sidekiq_client_middleware"
+require_relative "consistent_random/active_job"
 
 class ConsistentRandom
+  SEED_DIVISOR = (2**64 - 1).to_f
+  private_constant :SEED_DIVISOR
+
   class << self
     # Define a scope where consistent random values will be generated.
     #
+    # @param seeds [String, Symbol, Integer, Array<String>, Array<Integer>, nil] optional value to
+    #   use for generating random numbers. By default a random value will be generated. If the
+    #   scope is nested in another scope block, then the seed from the parent scope will be used
+    #   by default.
     # @yield block of code to execute within the scope
     # @return the result of the block
-    def scope
-      existing_context = Thread.current[:consistent_random_context]
+    def scope(seed = nil)
+      existing_seed = Thread.current[:consistent_random_seed]
+      seed_value = case seed
+      when nil
+        existing_seed || SecureRandom.hex
+      when String, Symbol, Integer
+        seed.to_s
+      when Array
+        seed.map { |s| s.to_s }.join("\x1C")
+      else
+        raise ArgumentError, "Invalid seed value: #{seed.inspect}"
+      end
+
       begin
-        context = Context.new(existing_context)
-        Thread.current[:consistent_random_context] = context
+        Thread.current[:consistent_random_seed] = seed_value
         yield
       ensure
-        Thread.current[:consistent_random_context] = existing_context
+        Thread.current[:consistent_random_seed] = existing_seed
       end
+    end
+
+    # Get the current seed used to generate random numbers. This will return nil if called
+    # outside of a scope.
+    #
+    # @return [String, nil] the seed value for the current scope
+    #  or nil if called outside of a scope.
+    # @api private
+    def current_seed
+      Thread.current[:consistent_random_seed]
     end
   end
 
@@ -27,7 +58,9 @@ class ConsistentRandom
     @name = name
   end
 
-  # Generate a random float. This method works the same as Kernel#rand.
+  # Generate a random number. The same number will be generated within a scope block.
+  # This method works the same as Kernel#rand. It will generate a consistent value even
+  # across Ruby versions and platforms.
   #
   # @param max [Integer, Range] the maximum value of the random float or a range indicating
   #  the minimum and maximum values.
@@ -35,35 +68,73 @@ class ConsistentRandom
   #   a number in that range. If max is an number, then it will be an integer between 0 and that
   #   value. Otherwise, it will be a float between 0 and 1.
   def rand(max = nil)
-    random.rand(max || 1.0)
+    value = seed / SEED_DIVISOR
+    case max
+    when nil
+      value
+    when Numeric
+      (value * max.to_i).to_i
+    when Range
+      cap_to_range(value, max)
+    end
   end
 
-  # Generate a random integer. This method works the same as Random#bytes.
+  # Generate a random array of bytes. The same number will be generated within a scope block.
+  # This method works the same as Random#bytes.
   #
   # @param size [Integer] the number of bytes to generate.
   # @return [String] a string of random bytes.
   def bytes(size)
-    random.bytes(size)
+    bytes = []
+    ((size + 19) / 20).times { |i| bytes << seed_hash("#{@name}#{i}").to_s }
+    bytes.join[0, size]
   end
 
-  # Generate a random number generator for the given name. The generator will always
-  # have the same seed within a scope.
+  # Generate a seed that can be used to generate random numbers. This seed will be
+  # return a consistent value when called within a scope.
   #
-  # @return [Random] a random number generator
-  def random
-    Random.new(current_context.seed(@name))
+  # @return [Integer] a 64 bit integer for seeding random values
+  def seed
+    hash = seed_hash(@name)
+    hash.byteslice(0, 8).unpack1("Q>")
   end
 
   # @return [Boolean] true if the other object is a ConsistentRandom that returns
   #   the same random number generator. If called outside of a scope, then it will
   #   always return false.
   def ==(other)
-    other.is_a?(self.class) && other.random = random
+    other.is_a?(self.class) && other.seed = seed
+  end
+
+  # Generate a random number generator for the given name. The generator will always
+  # have the same seed within a scope.
+  #
+  # This value is dependent on the Ruby Random class and may not generate consistent values
+  # across Ruby versions and platforms.
+  #
+  # @return [Random] a random number generator
+  def random
+    Random.new(seed)
   end
 
   private
 
-  def current_context
-    Thread.current[:consistent_random_context] || Context.new
+  def seed_hash(name)
+    random_seed = self.class.current_seed || SecureRandom.hex
+    Digest::SHA1.digest("#{random_seed}\x1C#{name}")
+  end
+
+  def cap_to_range(value, range)
+    min = range.begin
+    max = range.end
+    if min.nil? || max.nil?
+      raise ArgumentError, "Cannot generate random value for infinite range"
+    end
+
+    int_range = min.is_a?(Integer) && max.is_a?(Integer)
+    max += 1 if int_range && range.include?(max)
+
+    val = (value * (max - min)) + min
+    int_range ? val.to_i : val
   end
 end